First commit for merge of Grimaldi's backend and OmanAnwar's frontend.
We started out with a mirror of grimaldi.
We kept the grimaldi plant-list.service available because that informs us of what endpoints are currently available on the backend.
We kept administration folder from grimaldi.
We copied over the .html and .ts client-side of OmanAnwar and modified app.routes and app.module to recognize these new components.
We did not copy over the home component from OmanAnwar.

We have not begun refactoring OmanAnwar's html/ts to point to grimaldi's endpoints.