Mount bed list to the right of the page
This will work in place of the filter side bar we are currently building.
The bootstrap "container" class within app.component.html was overriding our css to have margins.
By removing the "container" class we removed the margins allowing the bed list to be anchored on the right
side of the page.

Addresses issues #4, #5