Implement rudimentary filter side bar
The interface is currently in the process of being overhauled. However, all functionality required has now been shown to work.
* Can have a vertical side bar with rotated text (this has been moved to the left side instead of the right as it is much easier and more reliable for a left side rotated nav bar)
* Can add elements to the vertical side bar
* Can tap/click a sidebar element to open an overlay
* Overlay's can have components embedded within them
* Overlay's can be closed once finished (auto-close in future)

It is noteworthy that this is obviously not a complete, or close to complete, interface. However, what is fantastic is that what we would like to have in the future already exists but in a crude form.

Addresses issues #4, #5, #7