Finish implementing bed page visit and QR bed page visit counts.
Two endpoints exist on the server: /api/bedVisit and /api/qrVisit

When ever someone chooses a bed in the dropdown a POST request is sent to /api/bedVisit containing which bed
that was incremented. The pageViews are incremented and a new visit is pushed on the bedVisits array.

When someone goes to a bed page manually on the browser that POST request is sent to /api/bedVisit. If the url contains
qr=true then qrScans is incremented and a new visit is pushed on the qrVisits array.